+++ changelog.md
@@ -1,4 +1,9 @@
 # Darko Changelog
 
+### *2019.07.09* - 1.0.1
+* Added optional "Load more posts" button
+* Changed related posts spinner
+* Fixed sticky post styles in blog
+
 ### *2019.06.30* - 1.0.0
 * Initial release